+++ src/Raven.CodeAnalysis/Syntax/TrailingBlockExpressionSyntax.cs
@@ -1,0 +1,6 @@
+namespace Raven.CodeAnalysis.Syntax;
+
+public sealed partial class TrailingBlockExpressionSyntax
+{
+    public BlockSyntax Body => SyntaxFactory.Block(OpenBraceToken, Statements, CloseBraceToken);
+}